Move old API endpoints to a new URL path
