+++ webserver/views/test/test_api_legacy.py
@@ -6,16 +6,6 @@ import os
 
 class APILegacyViewsTestCase(ServerTestCase):
 
-    def test_get_low_level(self):
-        mbid = '0dad432b-16cc-4bf0-8961-fd31d124b01b'
-        resp = self.client.get("/%s/low-level" % mbid)
-        self.assertEqual(resp.status_code, 404)
-
-        self.load_low_level_data(mbid)
-
-        resp = self.client.get("/%s/low-level" % mbid)
-        self.assertEqual(resp.status_code, 200)
-
     def test_submit_low_level(self):
         mbid = '0dad432b-16cc-4bf0-8961-fd31d124b01b'
 
@@ -26,14 +16,6 @@ class APILegacyViewsTestCase(ServerTestCase):
                                        content_type='application/json')
                 self.assertEqual(sub_resp.status_code, 200)
 
-        resp = self.client.get("/%s/low-level" % mbid)
+        # Getting from the new API
+        resp = self.client.get("/api/v1/%s/low-level" % mbid)
         self.assertEqual(resp.status_code, 200)
-
-    def test_cors_headers(self):
-        mbid = '0dad432b-16cc-4bf0-8961-fd31d124b01b'
-        self.load_low_level_data(mbid)
-
-        resp = self.client.get("/%s/low-level" % mbid)
-        self.assertEqual(resp.headers['Access-Control-Allow-Origin'], '*')
-
-        # TODO: Test in get_high_level.
